+++ CHANGES.rst
@@ -10,6 +10,7 @@ Version 0.9.0 (UNRELEASED)
 - Adds new ``/api/launch`` endpoint that allows running workflows from remote sources.
 - Adds the possibility to fetch the workflow specification from a remote URL.
 - Adds REANA specification validation utilities.
+- Adds `retention-rules-apply` command that can be used to apply pending retention rules.
 - Changes workflow create endpoint to populate workspace retention rules for the workflow.
 - Changes workflow list endpoint to return workspace retention rules for the workflow.
 - Changes API rate limiter error messages to be more verbose.